+++ applications/fishing-map/src/features/dataviews/dataviews.utils.ts
@@ -20,6 +20,7 @@ export const ENCOUNTER_EVENTS_SOURCE_ID = 'encounter-events'
 export const VESSEL_LAYER_PREFIX = 'vessel-'
 export const ENVIRONMENTAL_LAYER_PREFIX = 'environment-'
 export const CONTEXT_LAYER_PREFIX = 'context-'
+export const PRESENCE_LAYER_ID = 'presence'
 
 type VesselInstanceDatasets = {
   trackDatasetId: string
@@ -64,7 +65,7 @@ export const getFishingDataviewInstance = (): DataviewInstance<Generators.Type> 
 
 export const getPresenceDataviewInstance = (): DataviewInstance<Generators.Type> => {
   return {
-    id: `presence-${Date.now()}`,
+    id: `${PRESENCE_LAYER_ID}-${Date.now()}`,
     config: {
       colorCyclingType: 'fill' as ColorCyclingType,
     },